+++ src/components/BlogInput.js
@@ -1,0 +1,21 @@
+const BlogInput = ({ newBlog, setNewBlog, type }) => {
+  const handleForm = (e) => {
+    const newBlogObject = { ...newBlog }
+    newBlogObject[type] = e.target.value
+    setNewBlog(newBlogObject)
+  }
+
+  return (
+    <div>
+      <p>{type} :</p>
+      <input
+        type="text"
+        value={newBlog.type}
+        name={`Blog${type}`}
+        onChange={handleForm}
+      ></input>
+    </div>
+  )
+}
+
+export default BlogInput
